II-ESP-900/Adapt-IA_back/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
PyJWT
pydantic[email]
passlib[bcrypt]
requests
python-jose
python-multipart
black
pytest
httpx
python-decouple
cloudinary
pre-commit
pytest-asyncio
sqladmin
itsdangerous
pycryptodome
fastapi-mail
